Refactor dick measurement
