+++ src/Measurements.Core/Dick/Bones.cs
@@ -1,0 +1,14 @@
+namespace Measurements.Dick
+{
+    internal enum Bones
+    {
+        /// <summary>
+        /// Base of dick
+        /// </summary>
+        cm_J_dan100_00,
+        /// <summary>
+        /// Tip of dick
+        /// </summary>
+        cm_J_dan109_00,
+    }
+}
